fix: combined two dropdowns into a single component
